Move DataState into the weco namespace
