+++ common/internal_model/src/main/scala/uk/ac/wellcome/models/Implicits.scala
@@ -5,7 +5,7 @@ import io.circe._
 import uk.ac.wellcome.models.matcher.MatcherResult
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.models.work.internal._
-import weco.catalogue.internal_model.identifiers.{IdState, SourceIdentifier}
+import weco.catalogue.internal_model.identifiers._
 import weco.catalogue.internal_model.image.{
   Image,
   ImageData,
